#10 implement bookAppointment(), reserveAppointment(), getPatientInformation(), updatePatientInfo() for the New Appointment page
The issue countdown timer is related to a few other functionalities. In order to make the countdown feature work, the patient first needs to have the option to book the appointment and to book the appointment we need to get the patient's contact information in case the patient needs to update the contact info for this particular appointment and for this we need an update patient method, hence the implementation the four methods in order to have a complete solution instead of a partial one.