#29 add .env file to client_patient to store urls.
There was a mismatch between urls used in the .vue frontend files and the topics used in the backend. This blocked all requests from the frontend. While changing these urls I realized that we have so many files now and hardcoded urls are a bad practice as they make maintenance difficult and time consuming. These changes are not mentioned explicitly as a part of this issue because they should have been built in from the start. Now the .env file contains a few urls (not all) and we can add more as we go. Any changes to url now happens in one place. Right now, the urls are changed in the following .vue files: NewAppointment, MyAppointments, SingleAppointment, Profile, and MedicalJournal